Adds a Lookup() func
