+++ pkg/featflag/features.go
@@ -1,5 +1,6 @@
 //
 // Copyright 2019 Chef Software, Inc.
+// Author: Salim Afiune <afiune@chef.io>
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -135,24 +136,6 @@ func ListAll() string {
 	}
 
 	return strings.Join(list, " ")
-}
-
-func GetFromKey(key string) (*Feature, bool) {
-	for _, feat := range featureFlags {
-		if feat.configKey == key {
-			return &feat, true
-		}
-	}
-	return nil, false
-}
-
-func GetFromEnv(name string) (*Feature, bool) {
-	for _, feat := range featureFlags {
-		if feat.envName == name {
-			return &feat, true
-		}
-	}
-	return nil, false
 }
 
 func (feat *Feature) String() string {
