+++ pkg/featflag/features_test.go
@@ -1,5 +1,6 @@
 //
 // Copyright 2019 Chef Software, Inc.
+// Author: Salim Afiune <afiune@chef.io>
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -33,66 +34,6 @@ func TestDefaultFeatureFlags(t *testing.T) {
 		subject.ListAll(),
 		"default list of features doesn't match",
 	)
-}
-
-func TestGetFromEnvUnknown(t *testing.T) {
-	featFlag, exist := subject.GetFromEnv("CHEF_FEAT_BAR")
-	if assert.False(t, exist, "unknown feature flag should not exist") {
-		assert.Nil(t, featFlag, "unknown feature flag should return nil")
-	}
-}
-
-func TestGetFromEnvGlobalChefFeatAll(t *testing.T) {
-	featFlag, exist := subject.GetFromEnv("CHEF_FEAT_ALL")
-	if assert.True(t, exist, "global CHEF_FEAT_ALL:all flag should exist") {
-		assert.True(t,
-			subject.ChefFeatAll.Equals(featFlag),
-			"the feature flags don't match",
-		)
-		assert.False(t,
-			featFlag.Enabled(),
-			"global feature flag for all features should be disabled",
-		)
-		assert.Equal(t,
-			"CHEF_FEAT_ALL", featFlag.Env(),
-			"environment variable mismatch")
-		assert.Equal(t,
-			"all", featFlag.Key(),
-			"config key mismatch")
-		assert.Equal(t,
-			"(all:CHEF_FEAT_ALL)", subject.ChefFeatAll.String(),
-			"string conversion mismatch")
-	}
-}
-
-func TestGetFromKeyUnknown(t *testing.T) {
-	featFlag, exist := subject.GetFromKey("bar")
-	if assert.False(t, exist, "unknown feature flag should not exist") {
-		assert.Nil(t, featFlag, "unknown feature flag should return nil")
-	}
-}
-
-func TestGetFromKeyGlobalChefFeatAll(t *testing.T) {
-	featFlag, exist := subject.GetFromKey("all")
-	if assert.True(t, exist, "global CHEF_FEAT_ALL:all flag should exist") {
-		assert.True(t,
-			subject.ChefFeatAll.Equals(featFlag),
-			"the feature flags don't match",
-		)
-		assert.False(t,
-			featFlag.Enabled(),
-			"global feature flag for all features should be disabled",
-		)
-		assert.Equal(t,
-			"CHEF_FEAT_ALL", featFlag.Env(),
-			"environment variable mismatch")
-		assert.Equal(t,
-			"all", featFlag.Key(),
-			"config key mismatch")
-		assert.Equal(t,
-			"(all:CHEF_FEAT_ALL)", subject.ChefFeatAll.String(),
-			"string conversion mismatch")
-	}
 }
 
 func TestAllFeatureFlagsDisabledByDefault(t *testing.T) {
